Update tests with new strategy value
Use pydantic dataclasses instead of the built-in dataclasses.
Use `Any` instead of `dlite.Collection` for the type for stored DLite
Collections. This is due to pydantic not being able to validate
arbitrary types - and there is no way to allow this for a pydantic
dataclass. Then it needs to be converted to a pydantic BaseModel, but we
cannot do this.